Refactor category view and add custom table view

+++ src/main/java/com/r/projektnizad/controllers/category/View.java
@@ -2,29 +2,32 @@ package com.r.projektnizad.controllers.category;
 
 import com.r.projektnizad.dao.CategoryDao;
 import com.r.projektnizad.models.Category;
-import com.r.projektnizad.models.StoppableScene;
+import com.r.projektnizad.models.CleanableScene;
+import com.r.projektnizad.models.history.AddChange;
+import com.r.projektnizad.models.history.DeleteChange;
+import com.r.projektnizad.models.history.ModifyChange;
+import com.r.projektnizad.threads.ChangeWriterThread;
 import com.r.projektnizad.threads.SignaledTaskThread;
-import com.r.projektnizad.util.AppDialog;
-import com.r.projektnizad.util.CustomButtonTypes;
-import com.r.projektnizad.util.Navigator;
+import com.r.projektnizad.util.*;
 import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.*;
-import javafx.scene.input.MouseButton;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableRow;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.function.Consumer;
 
-public class View implements StoppableScene {
+public class View implements CleanableScene {
   public TableColumn<Category, String> descriptionTableColumn;
   public TableColumn<Category, String> nameTableColumn;
   public TableColumn<Category, Long> idTableColumn;
-
   private final CategoryDao categoryDao = new CategoryDao();
-  public TableView<Category> categoryTableView;
-
+  public CustomTableView<Category> categoryTableView;
   private final SignaledTaskThread<List<Category>> signaledTaskThread = new SignaledTaskThread<>((v) -> categoryDao.getAll());
 
   @FXML
@@ -34,36 +37,19 @@ public class View implements StoppableScene {
     nameTableColumn.setCellValueFactory(cellData -> new SimpleStringProperty(cellData.getValue().getName()));
     descriptionTableColumn.setCellValueFactory(cellData -> new SimpleStringProperty(cellData.getValue().getDescription()));
 
-    // add right click menu for delete and edit
+    Map<String, Consumer<Category>> actions = Map.of(
+            "Izmjeni", this::edit,
+            "Obriši", this::delete
+    );
+
     categoryTableView.setRowFactory(tableView -> {
-      var row = new TableRow<Category>();
-      row.setOnMouseClicked(event -> {
-        if (event.getButton() == MouseButton.SECONDARY && !row.isEmpty()) {
-          // edit and delete items
-          ContextMenu contextMenu = new ContextMenu();
-          var editItem = new MenuItem("Izmjeni");
-          var deleteItem = new MenuItem("Obriši");
-
-          editItem.setOnAction(event1 -> edit(row.getItem()));
-          deleteItem.setOnAction(event1 -> {
-            var confirm = new AppDialog().showConfirmationMessage("Obriši kategoriju", "Da li ste sigurni da želite obrisati kategoriju?", CustomButtonTypes.DELETE);
-            if (confirm == CustomButtonTypes.DELETE) {
-              delete(row.getItem());
-            }
-          });
-
-          contextMenu.getItems().addAll(editItem, deleteItem);
-          row.setContextMenu(contextMenu);
-        }
-      });
+      TableRow<Category> row = new TableRow<>();
+      TableViewContextMenu.build(row, actions);
       return row;
     });
 
-
-    // update with new data, update existing data
     signaledTaskThread.resultProperty().addListener((observable, oldValue, newValue) -> {
-      categoryTableView.getItems().clear();
-      categoryTableView.getItems().addAll(newValue);
+      categoryTableView.setItems(newValue);
     });
 
     signaledTaskThread.signal();
@@ -73,25 +59,30 @@ public class View implements StoppableScene {
     var editDialog = new AddDialog(Optional.ofNullable(category));
     editDialog.showAndWait().ifPresentOrElse(newCategory -> {
       categoryDao.update(category.getId(), newCategory);
+      new ChangeWriterThread<>(new ModifyChange<>(category, newCategory)).start();
       signaledTaskThread.signal();
     }, editDialog::close);
   }
 
   private void delete(Category category) {
+    ButtonType confirm = new AppDialog().showConfirmationMessage("Obriši kategoriju", "Da li ste sigurni da želite obrisati kategoriju?", CustomButtonTypes.DELETE);
+    if (confirm == CustomButtonTypes.CANCEL) return;
+
     categoryDao.delete(category.getId());
+    new ChangeWriterThread<>(new DeleteChange<>(category)).start();
     signaledTaskThread.signal();
   }
 
   @Override
-  public void stop() {
+  public void cleanup() {
     signaledTaskThread.interrupt();
   }
 
   public void openAddCategory(ActionEvent actionEvent) {
     var addDialog = new AddDialog(Optional.empty());
-
     addDialog.showAndWait().ifPresent(category -> {
       categoryDao.save(category);
+      new ChangeWriterThread<>(new AddChange<>(category)).start();
       signaledTaskThread.signal();
     });
